requirements.txt
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
click==7.1.1
coverage==5.0.4
fastapi==0.54.1
h11==0.9.0
httptools==0.1.1
idna==2.9
more-itertools==8.2.0
packaging==20.3
passlib==1.7.2
pluggy==0.13.1
psycopg2==2.8.5
py==1.8.1
pycparser==2.20
pydantic==1.4
PyJWT==1.7.1
pyparsing==2.4.7
pytest==5.4.1
requests==2.23.0
six==1.14.0
SQLAlchemy==1.3.16
starlette==0.13.2
urllib3==1.25.8
uvicorn==0.11.3
uvloop==0.14.0
wcwidth==0.1.9
websockets==8.1
